Enable Rack integration for Sinatra
This attemps to mimic what is currently done with Rails.

Sinatra provides only limited public facilities to manipulate the stack
(namely, appending with `use`), assumes his default middlewares are
always first, and has a peculiar way to wrap the middlewares to build
the app stack.

Sinatra also freezes the stack when the builder is run, so we have to
get access to the moment the builder is available but not yet run to
alter the list by patching `Sinatra::Base#setup_default_middlewares`.

The `Datadog::Contrib::Sinatra::Framework` class encapsulates these
internal implementation details behind a few methods to achieve the
desired manipulations.